refactor: move state Screenshots to new file
