+++ src/templates/state.js
@@ -1,46 +1,16 @@
 import React from 'react'
 import { graphql } from 'gatsby'
-import { DateTime } from 'luxon'
 import marked from 'marked'
 import Layout from '../components/layout'
 import formatDate from '../utilities/format-date'
 import thousands from '../utilities/format-thousands'
+import Screenshots from '../components/pages/state/screenshots'
 import StateGrade from '../components/common/state-grade'
 import StateLinks from '../components/pages/state/state-links'
 import SummaryTable from '../components/common/summary-table'
 import { SyncInfobox } from '../components/common/infobox'
 import Table from '../components/common/table'
 import '../scss/templates/state.scss'
-
-const Screenshots = ({ date, screenshots }) => {
-  const dateScreenshots = []
-  const currentDate = DateTime.fromISO(date)
-  screenshots.forEach(({ node }) => {
-    if (DateTime.fromISO(node.dateChecked).hasSame(currentDate, 'day')) {
-      dateScreenshots.push(node)
-    }
-  })
-  if (dateScreenshots.length === 0) {
-    return null
-  }
-  return (
-    <ul>
-      {dateScreenshots.map(screenshot => (
-        <li key={screenshot.url}>
-          <a href={screenshot.url} target="_blank" rel="noopener noreferrer">
-            {screenshot.dateChecked && (
-              <>
-                {DateTime.fromISO(screenshot.dateChecked)
-                  .toFormat('h:mm a')
-                  .toLowerCase()}
-              </>
-            )}
-          </a>
-        </li>
-      ))}
-    </ul>
-  )
-}
 
 const StateHistory = ({ history, screenshots }) => (
   <Table className="state-historical">
